+++ src/features/Home/index.js
@@ -4,7 +4,6 @@ import Head from 'common/components/Head';
 import Container from 'common/components/Container';
 
 import styles from './styles.less';
-import News from './components/News';
 import Introduction from './components/Introduction';
 import RandomCharacter from './components/RandomCharacter';
 import RandomGuilds from './components/RandomGuilds';
@@ -21,12 +20,6 @@ const Home = () => (
     </div>
 
     <RandomGuilds />
-
-    <Container>
-      <hr />
-
-      <News className={styles.newsContainer} />
-    </Container>
   </div>
 );